Remove hungarian notation from class names
