--- app/src/main/java/com/esoxjem/movieguide/details/IMovieDetailsInteractor.java
@@ -1,17 +1,0 @@
-package com.esoxjem.movieguide.details;
-
-import com.esoxjem.movieguide.Review;
-import com.esoxjem.movieguide.Video;
-
-import java.util.List;
-
-import rx.Observable;
-
-/**
- * @author arun
- */
-public interface IMovieDetailsInteractor
-{
-    Observable<List<Video>> getTrailers(String id);
-    Observable<List<Review>> getReviews(String id);
-}
