+++ app/src/main/java/com/esoxjem/movieguide/details/MovieDetailsPresenter.java
@@ -1,163 +1,23 @@
 package com.esoxjem.movieguide.details;
 
 import com.esoxjem.movieguide.Movie;
-import com.esoxjem.movieguide.Review;
-import com.esoxjem.movieguide.Video;
-import com.esoxjem.movieguide.favorites.IFavoritesInteractor;
-import com.esoxjem.movieguide.util.RxUtils;
-
-import java.util.List;
-
-import rx.Subscriber;
-import rx.Subscription;
-import rx.android.schedulers.AndroidSchedulers;
-import rx.schedulers.Schedulers;
 
 /**
  * @author arun
  */
-class MovieDetailsPresenter implements IMovieDetailsPresenter
+public interface MovieDetailsPresenter
 {
-    private IMovieDetailsView view;
-    private IMovieDetailsInteractor movieDetailsInteractor;
-    private IFavoritesInteractor favoritesInteractor;
-    private Subscription trailersSubscription;
-    private Subscription reviewSubscription;
+    void showDetails(Movie movie);
 
-    MovieDetailsPresenter(IMovieDetailsInteractor movieDetailsInteractor, IFavoritesInteractor favoritesInteractor)
-    {
-        this.movieDetailsInteractor = movieDetailsInteractor;
-        this.favoritesInteractor = favoritesInteractor;
-    }
+    void showTrailers(Movie movie);
 
-    @Override
-    public void setView(IMovieDetailsView view)
-    {
-        this.view = view;
-    }
+    void showReviews(Movie movie);
 
-    @Override
-    public void destroy()
-    {
-        view = null;
-        RxUtils.unsubscribe(trailersSubscription, reviewSubscription);
-    }
+    void showFavoriteButton(Movie movie);
 
-    @Override
-    public void showDetails(Movie movie)
-    {
-        if (isViewAttached())
-        {
-            view.showDetails(movie);
-        }
-    }
+    void onFavoriteClick(Movie movie);
 
-    private boolean isViewAttached()
-    {
-        return view != null;
-    }
+    void setView(MovieDetailsView view);
 
-    @Override
-    public void showTrailers(Movie movie)
-    {
-        trailersSubscription = movieDetailsInteractor.getTrailers(movie.getId()).subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(new Subscriber<List<Video>>()
-                {
-                    @Override
-                    public void onCompleted()
-                    {
-                        // do nothing
-                    }
-
-                    @Override
-                    public void onError(Throwable e)
-                    {
-                        // do nothing
-                    }
-
-                    @Override
-                    public void onNext(List<Video> videos)
-                    {
-                        onGetTrailersSuccess(videos);
-                    }
-                });
-    }
-
-    private void onGetTrailersSuccess(List<Video> videos)
-    {
-        if (isViewAttached())
-        {
-            view.showTrailers(videos);
-        }
-    }
-
-    @Override
-    public void showReviews(Movie movie)
-    {
-        reviewSubscription = movieDetailsInteractor.getReviews(movie.getId()).subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(new Subscriber<List<Review>>()
-                {
-                    @Override
-                    public void onCompleted()
-                    {
-                        // do nothing
-                    }
-
-                    @Override
-                    public void onError(Throwable e)
-                    {
-                        // do nothing
-                    }
-
-                    @Override
-                    public void onNext(List<Review> reviews)
-                    {
-                        onGetReviewsSuccess(reviews);
-                    }
-                });
-    }
-
-    private void onGetReviewsSuccess(List<Review> reviews)
-    {
-        if (isViewAttached())
-        {
-            view.showReviews(reviews);
-        }
-    }
-
-    @Override
-    public void showFavoriteButton(Movie movie)
-    {
-        boolean isFavorite = favoritesInteractor.isFavorite(movie.getId());
-        if (isViewAttached())
-        {
-            if (isFavorite)
-            {
-                view.showFavorited();
-            } else
-            {
-                view.showUnFavorited();
-            }
-        }
-    }
-
-    @Override
-    public void onFavoriteClick(Movie movie)
-    {
-        if (isViewAttached())
-        {
-            boolean isFavorite = favoritesInteractor.isFavorite(movie.getId());
-            if (isFavorite)
-            {
-                favoritesInteractor.unFavorite(movie.getId());
-                view.showUnFavorited();
-            } else
-            {
-                favoritesInteractor.setFavorite(movie);
-                view.showFavorited();
-            }
-        }
-    }
+    void destroy();
 }
