+++ app/src/main/java/com/esoxjem/movieguide/details/MovieDetailsView.java
@@ -9,7 +9,7 @@ import java.util.List;
 /**
  * @author arun
  */
-public interface IMovieDetailsView
+interface MovieDetailsView
 {
     void showDetails(Movie movie);
     void showTrailers(List<Video> trailers);
