+++ app/src/main/java/com/esoxjem/movieguide/favorites/FavoritesInteractor.java
@@ -2,49 +2,15 @@ package com.esoxjem.movieguide.favorites;
 
 import com.esoxjem.movieguide.Movie;
 
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
 /**
  * @author arun
  */
-public class FavoritesInteractor implements IFavoritesInteractor
+public interface FavoritesInteractor
 {
-    private FavoritesStore favoritesStore;
-
-    public FavoritesInteractor(FavoritesStore store)
-    {
-        favoritesStore = store;
-    }
-
-    @Override
-    public void setFavorite(Movie movie)
-    {
-        favoritesStore.setFavorite(movie);
-    }
-
-    @Override
-    public boolean isFavorite(String id)
-    {
-        return favoritesStore.isFavorite(id);
-    }
-
-    @Override
-    public List<Movie> getFavorites()
-    {
-        try
-        {
-            return favoritesStore.getFavorites();
-        } catch (IOException ignored)
-        {
-            return new ArrayList<>(0);
-        }
-    }
-
-    @Override
-    public void unFavorite(String id)
-    {
-        favoritesStore.unfavorite(id);
-    }
+    void setFavorite(Movie movie);
+    boolean isFavorite(String id);
+    List<Movie> getFavorites();
+    void unFavorite(String id);
 }
