--- app/src/main/java/com/esoxjem/movieguide/favorites/IFavoritesInteractor.java
@@ -1,16 +1,0 @@
-package com.esoxjem.movieguide.favorites;
-
-import com.esoxjem.movieguide.Movie;
-
-import java.util.List;
-
-/**
- * @author arun
- */
-public interface IFavoritesInteractor
-{
-    void setFavorite(Movie movie);
-    boolean isFavorite(String id);
-    List<Movie> getFavorites();
-    void unFavorite(String id);
-}
