--- app/src/main/java/com/esoxjem/movieguide/listing/IMoviesListingInteractor.java
@@ -1,15 +1,0 @@
-package com.esoxjem.movieguide.listing;
-
-import com.esoxjem.movieguide.Movie;
-
-import java.util.List;
-
-import rx.Observable;
-
-/**
- * @author arun
- */
-public interface IMoviesListingInteractor
-{
-    Observable<List<Movie>> fetchMovies();
-}
