--- app/src/main/java/com/esoxjem/movieguide/listing/IMoviesListingPresenter.java
@@ -1,13 +1,0 @@
-package com.esoxjem.movieguide.listing;
-
-/**
- * @author arun
- */
-public interface IMoviesListingPresenter
-{
-    void displayMovies();
-
-    void setView(IMoviesListingView view);
-
-    void destroy();
-}
