+++ app/src/main/java/com/esoxjem/movieguide/listing/MoviesListingView.java
@@ -7,7 +7,7 @@ import java.util.List;
 /**
  * @author arun
  */
-public interface IMoviesListingView
+interface MoviesListingView
 {
     void showMovies(List<Movie> movies);
     void loadingStarted();
