--- app/src/main/java/com/esoxjem/movieguide/listing/sorting/ISortingDialogInteractor.java
@@ -1,11 +1,0 @@
-package com.esoxjem.movieguide.listing.sorting;
-
-/**
- * @author arun
- */
-public interface ISortingDialogInteractor
-{
-    int getSelectedSortingOption();
-
-    void setSortingOption(SortType sortType);
-}
