--- app/src/main/java/com/esoxjem/movieguide/listing/sorting/ISortingDialogPresenter.java
@@ -1,19 +1,0 @@
-package com.esoxjem.movieguide.listing.sorting;
-
-/**
- * @author arun
- */
-public interface ISortingDialogPresenter
-{
-    void setLastSavedOption();
-
-    void onPopularMoviesSelected();
-
-    void onHighestRatedMoviesSelected();
-
-    void onFavoritesSelected();
-
-    void setView(ISortingDialogView view);
-
-    void destroy();
-}
