+++ app/src/main/java/com/esoxjem/movieguide/listing/sorting/SortingDialogPresenter.java
@@ -3,80 +3,17 @@ package com.esoxjem.movieguide.listing.sorting;
 /**
  * @author arun
  */
-public class SortingDialogPresenter implements ISortingDialogPresenter
+public interface SortingDialogPresenter
 {
-    private ISortingDialogView view;
-    private ISortingDialogInteractor sortingDialogInteractor;
+    void setLastSavedOption();
 
-    public SortingDialogPresenter(ISortingDialogInteractor interactor)
-    {
-        sortingDialogInteractor = interactor;
-    }
+    void onPopularMoviesSelected();
 
-    @Override
-    public void setView(ISortingDialogView view)
-    {
-        this.view = view;
-    }
+    void onHighestRatedMoviesSelected();
 
-    @Override
-    public void destroy()
-    {
-        view = null;
-    }
+    void onFavoritesSelected();
 
-    @Override
-    public void setLastSavedOption()
-    {
-        if (isViewAttached())
-        {
-            int selectedOption = sortingDialogInteractor.getSelectedSortingOption();
+    void setView(SortingDialogView view);
 
-            if (selectedOption == SortType.MOST_POPULAR.getValue())
-            {
-                view.setPopularChecked();
-            } else if (selectedOption == SortType.HIGHEST_RATED.getValue())
-            {
-                view.setHighestRatedChecked();
-            } else
-            {
-                view.setFavoritesChecked();
-            }
-        }
-    }
-
-    private boolean isViewAttached()
-    {
-        return view != null;
-    }
-
-    @Override
-    public void onPopularMoviesSelected()
-    {
-        if (isViewAttached())
-        {
-            sortingDialogInteractor.setSortingOption(SortType.MOST_POPULAR);
-            view.dismissDialog();
-        }
-    }
-
-    @Override
-    public void onHighestRatedMoviesSelected()
-    {
-        if (isViewAttached())
-        {
-            sortingDialogInteractor.setSortingOption(SortType.HIGHEST_RATED);
-            view.dismissDialog();
-        }
-    }
-
-    @Override
-    public void onFavoritesSelected()
-    {
-        if (isViewAttached())
-        {
-            sortingDialogInteractor.setSortingOption(SortType.FAVORITES);
-            view.dismissDialog();
-        }
-    }
+    void destroy();
 }
